requirements.txt
streamlit
google-generativeai
PyMuPDF
Pillow
pydantic
